+++ task_2/task2_2_3.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 int main(){
-    double x, a, result, deriative, integral;
+    float x, a, result, deriative, integral;
 	int n;
     printf("Введите степень многочлена\n");
 	scanf("%d", &n);
